finish chopping up
